+++ ParcelKit/PKSyncManager.m
@@ -28,10 +28,12 @@
 #import "DBRecord+ParcelKit.h"
 
 NSString * const PKDefaultSyncAttributeName = @"syncID";
+NSString * const PKSyncManagerManagedObjectContextIdentifier = @"PKSyncManagerPrivateManagedObjectContext";
 NSString * const PKSyncManagerDatastoreStatusDidChangeNotification = @"PKSyncManagerDatastoreStatusDidChange";
 NSString * const PKSyncManagerDatastoreStatusKey = @"status";
 NSString * const PKSyncManagerDatastoreIncomingChangesNotification = @"PKSyncManagerDatastoreIncomingChanges";
 NSString * const PKSyncManagerDatastoreIncomingChangesKey = @"changes";
+NSString * const PKSyncManagerDatastoreIncomingChangesErrorKey = @"error";
 NSString * const PKSyncManagerDatastoreLastSyncDateNotification = @"PKSyncManagerDatastoreLastSyncDateNotification";
 NSString * const PKSyncManagerDatastoreLastSyncDateKey = @"lastSyncDate";
 
@@ -52,7 +54,7 @@ NSString * const PKSyncManagerDatastoreLastSyncDateKey = @"lastSyncDate";
     return [uuid stringByReplacingOccurrencesOfString:@"-" withString:@""];
 }
 
-- (id)init
+- (instancetype)init
 {
     self = [super init];
     if (self) {
@@ -63,7 +65,7 @@ NSString * const PKSyncManagerDatastoreLastSyncDateKey = @"lastSyncDate";
     return self;
 }
 
-- (id)initWithManagedObjectContext:(NSManagedObjectContext *)managedObjectContext datastore:(DBDatastore *)datastore
+- (instancetype)initWithManagedObjectContext:(NSManagedObjectContext *)managedObjectContext datastore:(DBDatastore *)datastore
 {
     self = [self init];
     if (self) {
@@ -181,8 +183,10 @@ NSString * const PKSyncManagerDatastoreLastSyncDateKey = @"lastSyncDate";
 }
 
 #pragma mark - Updating Core Data
-- (BOOL)updateCoreDataWithDatastoreChanges:(NSDictionary *)changes
+- (BOOL)updateCoreDataWithDatastoreChanges:(NSDictionary *)changes error:(NSError **)error
 {
+    NSAssert(*error == nil, @"Error must be nil");
+
     static NSString * const PKUpdateManagedObjectKey = @"object";
     static NSString * const PKUpdateRecordKey = @"record";
     
@@ -190,7 +194,9 @@ NSString * const PKSyncManagerDatastoreLastSyncDateKey = @"lastSyncDate";
     
     NSManagedObjectContext *managedObjectContext = [[NSManagedObjectContext alloc] initWithConcurrencyType:NSPrivateQueueConcurrencyType];
     [managedObjectContext setPersistentStoreCoordinator:self.persistentStoreCoordinator];
+    [managedObjectContext setMergePolicy:NSMergeByPropertyObjectTrumpMergePolicy];
     [managedObjectContext setUndoManager:nil];
+    managedObjectContext.userInfo[PKSyncManagerManagedObjectContextIdentifier] = [NSObject new];
 
     __weak typeof(self) weakSelf = self;
     [managedObjectContext performBlockAndWait:^{
@@ -210,8 +216,8 @@ NSString * const PKSyncManagerDatastoreLastSyncDateKey = @"lastSyncDate";
             for (DBRecord *record in records) {
                 [fetchRequest setPredicate:[NSPredicate predicateWithFormat:@"%K == %@", strongSelf.syncAttributeName, record.recordId]];
                 
-                NSError *error = nil;
-                NSArray *managedObjects = [managedObjectContext executeFetchRequest:fetchRequest error:&error];
+                NSError *fetchRequestError = nil;
+                NSArray *managedObjects = [managedObjectContext executeFetchRequest:fetchRequest error:&fetchRequestError];
                 if (managedObjects)  {
                     NSManagedObject *managedObject = [managedObjects lastObject];
                     
@@ -228,24 +234,30 @@ NSString * const PKSyncManagerDatastoreLastSyncDateKey = @"lastSyncDate";
                         [updates addObject:@{PKUpdateManagedObjectKey: managedObject, PKUpdateRecordKey: record}];
                     }
                 } else {
-                    NSLog(@"Error executing fetch request: %@", error);
+                    NSLog(@"Error executing fetch request: %@", fetchRequestError);
                 }
             }
         }];
         
-        
-        for (NSDictionary *update in updates) {
-            NSManagedObject *managedObject = update[PKUpdateManagedObjectKey];
-            DBRecord *record = update[PKUpdateRecordKey];
-            [managedObject pk_setPropertiesWithRecord:record syncAttributeName:strongSelf.syncAttributeName];
+        @try {
+            for (NSDictionary *update in updates) {
+                NSManagedObject *managedObject = update[PKUpdateManagedObjectKey];
+                DBRecord *record = update[PKUpdateRecordKey];
+                [managedObject pk_setPropertiesWithRecord:record syncAttributeName:strongSelf.syncAttributeName];
+            }
         }
-        
+        @catch (NSException *exception) {
+            NSMutableDictionary *errorUserInfo = [@{@"exceptionName" : exception.name, @"exceptionReason" : exception.reason} mutableCopy];
+            if (exception.userInfo != nil) {
+                errorUserInfo[@"exceptionUserInfo"] = exception.userInfo;
+            }
+            *error = [NSError errorWithDomain:@"ParcelKit" code:1 userInfo:errorUserInfo];
+            return;
+        }
+
         if ([managedObjectContext hasChanges]) {
             [[NSNotificationCenter defaultCenter] addObserver:strongSelf selector:@selector(syncManagedObjectContextDidSave:) name:NSManagedObjectContextDidSaveNotification object:managedObjectContext];
-            NSError *error = nil;
-            if (![managedObjectContext save:&error]) {
-                NSLog(@"Error saving managed object context: %@", error);
-            }
+            [managedObjectContext save:error];
             [[NSNotificationCenter defaultCenter] removeObserver:strongSelf name:NSManagedObjectContextDidSaveNotification object:managedObjectContext];
         }
     }];
@@ -265,10 +277,15 @@ NSString * const PKSyncManagerDatastoreLastSyncDateKey = @"lastSyncDate";
 #pragma mark - Updating Datastore
 - (void)managedObjectContextWillSave:(NSNotification *)notification
 {
-    if (![self isObserving]) return;
-    
     NSManagedObjectContext *managedObjectContext = notification.object;
     if (self.managedObjectContext != managedObjectContext) return;
+    
+    [self updateDatastoreWithManagedObjectContextChanges:managedObjectContext];
+}
+
+- (void)updateDatastoreWithManagedObjectContextChanges:(NSManagedObjectContext *)managedObjectContext
+{
+    if (![self isObserving]) return;
     
     NSSet *deletedObjects = [managedObjectContext deletedObjects];
     for (NSManagedObject *managedObject in [self syncableManagedObjectsFromManagedObjects:deletedObjects]) {
@@ -318,8 +335,13 @@ NSString * const PKSyncManagerDatastoreLastSyncDateKey = @"lastSyncDate";
     DBError *error = nil;
     NSDictionary *changes = [self.datastore sync:&error];
     if (changes) {
-        if ([self updateCoreDataWithDatastoreChanges:changes]) {
-            [[NSNotificationCenter defaultCenter] postNotificationName:PKSyncManagerDatastoreIncomingChangesNotification object:self userInfo:@{PKSyncManagerDatastoreIncomingChangesKey: changes}];
+        NSError *updateError = nil;
+        NSMutableDictionary *userInfo = [@{PKSyncManagerDatastoreIncomingChangesKey : changes} mutableCopy];
+        if ([self updateCoreDataWithDatastoreChanges:changes error:&updateError]) {
+            if (updateError != nil) {
+                userInfo[PKSyncManagerDatastoreIncomingChangesErrorKey] = updateError;
+            }
+            [[NSNotificationCenter defaultCenter] postNotificationName:PKSyncManagerDatastoreIncomingChangesNotification object:self userInfo:userInfo];
         }
         [[NSNotificationCenter defaultCenter] postNotificationName:PKSyncManagerDatastoreLastSyncDateNotification object:self userInfo:@{PKSyncManagerDatastoreLastSyncDateKey: [NSDate date]}];